+++ Packages/Sources/Core/Composables/NavigationCommand.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum NavigationCommand: Equatable {
+  case show
+  case hide
+}
