+++ Packages/Sources/Core/Octokit/README.md
@@ -1,0 +1,2 @@
+# Octokit
+## Swift client for the Github.com REST API.
