+++ Packages/Sources/Features/Login/Sources/LoginReducer.swift
@@ -1,29 +1,7 @@
 import ComposableArchitecture
+import Composables
 import Foundation
 import Octokit
-
-// TODO: extract this
-public enum NavigationCommand: Equatable {
-  case show
-  case hide
-}
-
-public protocol ComposableState: Equatable {}
-
-public protocol ComposableAction: Equatable {
-  associatedtype User
-  associatedtype Local
-  associatedtype Delegate
-
-  static func user(_: User) -> Self
-  static func local(_: Local) -> Self
-  static func delegate(_: Delegate) -> Self
-}
-
-public protocol ComposableReducer: ReducerProtocol where State: ComposableState, Action: ComposableAction {
-  associatedtype ComposableBody: ReducerProtocol<State, Action>
-  var body: ComposableBody { get }
-}
 
 public struct LoginReducer: ComposableReducer {
   public struct State: ComposableState {
